optimize shooting stars and add more
