+++ app/challenge/landing-page/page.tsx
@@ -5,7 +5,7 @@ import styles from "./LandingPage.module.scss";
 import { useParallaxScrollY } from "@/hooks/use-parallax-scrollY";
 import clsx from "clsx";
 import { motion, Variants } from "framer-motion";
-import { ShootingStars } from "@/components/ShootingStars/ShootingStars";
+import { ShootingStar } from "@/components/ShootingStars/ShootingStar";
 
 const ProChallenge: React.FC = () => {
     const { offsetY, ref } = useParallaxScrollY();
@@ -122,7 +122,9 @@ const ProChallenge: React.FC = () => {
                 />
 
                 {/* Shooting Stars */}
-                <ShootingStars />
+                <ShootingStar />
+                <ShootingStar />
+                <ShootingStar />
 
                 {/* OVERLAY LAYER */}
                 <Box sx={{ position: "absolute", inset: 0 }}>
