+++ app/landing/Hero.tsx
@@ -8,7 +8,7 @@ import Image from "next/image";
 import { START_END_DATES } from "./constants";
 import styles from "./Hero.module.scss";
 import { motion, Variants } from "framer-motion";
-import { ShootingStars } from "@/components/ShootingStars/ShootingStars";
+import { ShootingStar } from "@/components/ShootingStars/ShootingStar";
 
 // Create a motion-enabled version of the Next.js Image component
 const MotionImage = motion(Image);
@@ -106,7 +106,9 @@ const Hero = () => {
             </div>
 
             {/* Shooting Stars */}
-            <ShootingStars size={500} />
+            <ShootingStar />
+            <ShootingStar />
+            <ShootingStar />
 
             {/* 3. Apply the container variants to the main content wrapper */}
             <motion.div
